% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonym.R
\name{humanSyno}
\alias{humanSyno}
\title{Human wraper for \code{\link{geneSynonym}}}
\usage{
humanSyno(genes, cores = 1)
}
\arguments{
\item{genes}{character vector of genes}

\item{cores}{number of cores to use when multiprocessing. Useful for large gene lists}
}
\description{
Human wraper for \code{\link{geneSynonym}}
}
\examples{
geneSynonym(c('MOG','ENO2'), tax = 9606)
geneSynonym(c('MOG',2026), tax = 9606)

}
