% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonym.R
\name{mouseSyno}
\alias{mouseSyno}
\title{Mouse wraper for \code{\link{geneSynonym}}}
\usage{
mouseSyno(genes, caseSensitive = TRUE)
}
\arguments{
\item{genes}{character vector of genes}
}
\description{
Mouse wraper for \code{\link{geneSynonym}}
}
\examples{
mouseSyno(c('Eno2','Mog'))
mouseSyno(c('Eno2',17441))

}
