% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{syno10090}
\alias{syno10090}
\title{Synonym information for Mus musculus}
\format{An object of class \code{character} of length 0.}
\usage{
syno10090
}
\description{
Synonym information for Mus musculus
}
\keyword{datasets}
