% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{syno10116}
\alias{syno10116}
\title{Synonym information for Rattus norvegicus}
\format{
An object of class \code{character} of length 62192.
}
\usage{
syno10116
}
\description{
Synonym information for Rattus norvegicus
}
\keyword{datasets}
