% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{syno562}
\alias{syno562}
\title{Synonym information for Escherichia coli}
\format{An object of class \code{character} of length 14841.}
\usage{
syno562
}
\description{
Synonym information for Escherichia coli
}
\keyword{datasets}
