% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{syno6239}
\alias{syno6239}
\title{Synonym information for Caenorhabditis elegans}
\format{An object of class \code{character} of length 46847.}
\usage{
syno6239
}
\description{
Synonym information for Caenorhabditis elegans
}
\keyword{datasets}
