% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{syno7227}
\alias{syno7227}
\title{Synonym information for Drosophila melanogaster}
\format{An object of class \code{character} of length 25040.}
\usage{
syno7227
}
\description{
Synonym information for Drosophila melanogaster
}
\keyword{datasets}
