% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{syno9544}
\alias{syno9544}
\title{Synonym information for Macaca mulatta}
\format{An object of class \code{character} of length 40390.}
\usage{
syno9544
}
\description{
Synonym information for Macaca mulatta
}
\keyword{datasets}
