% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonym.R
\name{geneSynonym}
\alias{geneSynonym}
\title{Get synonyms of given genes}
\usage{
geneSynonym(genes, tax, caseSensitive = TRUE)
}
\arguments{
\item{genes}{character vector of genes}

\item{tax}{Species taxonomy ID}
}
\description{
Given a list of genes and taxid, returns a data frame with corresponding synonyms to said genes. For mouse and humans just use humanSyno and mouseSyno
}
\examples{
geneSynonym(c('Eno2','Mog'), tax = 10090)
geneSynonym(c('Eno2',17441), tax = 10090)
geneSynonym(c('MOG','ENO2'), tax = 9606)
}
