% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonym.R
\name{humanSyno}
\alias{humanSyno}
\title{Human wraper for \code{\link{geneSynonym}}}
\usage{
humanSyno(genes)
}
\arguments{
\item{genes}{character vector of genes}
}
\description{
Human wraper for \code{\link{geneSynonym}}
}
\examples{
humanSyno(c('MOG','ENO2'))
humanSyno(c('MOG',2026))

}
