% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonym.R
\name{mouseSyno}
\alias{mouseSyno}
\title{Mouse wraper for \code{\link{geneSynonym}}}
\usage{
mouseSyno(genes, cores = 1)
}
\arguments{
\item{genes}{character vector of genes}

\item{cores}{number of cores to use when multiprocessing. Useful for large gene lists}
}
\description{
Mouse wraper for \code{\link{geneSynonym}}
}
