% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{syno7955}
\alias{syno7955}
\title{Synonym information for Danio rerio}
\format{
An object of class \code{character} of length 54882.
}
\usage{
syno7955
}
\description{
Synonym information for Danio rerio
}
\keyword{datasets}
