% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{syno9606}
\alias{syno9606}
\title{Synonym information for Homo sapiens}
\format{
An object of class \code{character} of length 193838.
}
\usage{
syno9606
}
\description{
Synonym information for Homo sapiens
}
\keyword{datasets}
